int
int

+++ Combinations Calculator.py
@@ -1,6 +1,6 @@
 
-n = float(input())
-r = float(input())
+n = int(input())
+r = int(input())
 
 def factorial(number):
    if number < 1:   
@@ -8,7 +8,7 @@ def factorial(number):
    else:
        return number * factorial( number - 1 )
 nCr = factorial(n)/(factorial(r)*(factorial(n-r)))
-print(int(nCr))
+print(nCr)
 
 ##this is my combinations calculator. getting factorial recursively. really ghetto imo but does the job i guess
